\name{toJSONarray}
\alias{toJSONarray}
\title{Internal function from Wei Luo to convert a data frame to a JSON array}
\source{
  Function from:
  http://theweiluo.wordpress.com/2011/09/30/r-to-json-for-d3-js-and-protovis/
}
\usage{
  toJSONarray(dtf)
}
\arguments{
  \item{dtf}{a data frame object.}
}
\description{
  Internal function from Wei Luo to convert a data frame to
  a JSON array
}

